Added Invincibility - World gen spawns walls instead of Prefab
The way invincibility works is by disabling collision between default and AlwaysCollision layers. Walls and the ball spawn on AlwaysCollision layer . The parent of a prefab is on the AlwaysCollision layer so spawncollision can keep doing its work. Everything else is on the default layer so the collision will disable. This is a really cheap way to do it so it is subjected to change.